requirements.txt
discord.py
time
bs4
urllib

